requirements.txt
numpy>=2.1
sounddevice>=0.4.7
soundfile>=0.12
paho-mqtt>=1.6
matplotlib>=3.9
ttkbootstrap>=1.10
PyYAML>=6.0
pydub>=0.25
